app/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary  
alembic          
python-dotenv    
asyncpg          
prometheus-client==0.19.0
prometheus-fastapi-instrumentator

